Restructured project to use interfaces instead of packages
